app/templates/requirements/base.txt
Django==1.6.2
South==0.8.4
psycopg2==2.5.2
django-compressor==1.3
django-modelcluster==0.1
-e git://github.com/torchbox/wagtail.git#egg=wagtail
django-taggit==0.11.2
django-libsass==0.1
django_extensions
